CLN : literalinclude multi_object_view
	modified:   docs/source/traitsui_user_manual/advanced_view.rst
	new file:   docs/source/traitsui_user_manual/examples/multi_object_view.py
	deleted:    examples/tutorials/doc_examples/examples/multi_object_view.py
